+++ app/src/at-hieucap/AndroidManifest.xml
@@ -18,6 +18,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".calculator.CalculatorActivity"/>
-        <activity android:name=".recyclerView.PersonViewActivity"/>
+        <activity android:name=".recycler.view.PersonViewActivity"/>
     </application>
 </manifest>
